triggering the layout protocol
CHANGE-SPACE-REQUIREMENTS triggers the layout protocol only when the argument
:RESIZE-FRAME is T. Otherwise it calls NOTE-SPACE-REQUIREMENTS-CHANGED and it
is the parent who decides what to do.

Usually (NOTE-SPACE-REQUIREMENTS-CHANGED parent child)
calls (CHANGE-SPACE-REQUIREMENTS parent), until the message arrives to the
top-level-pane that triggers the layout protocol calling LAYOUT-FRAME.

Exceptions are RESTRAINING-PANE that doesn't do anything, and VIEWPORT-PANE
that doesn't change its space-requirements but allocates its space without
changing its dimensions in order to update which part of the child is viewed.