map-over-command-table-*: provide uniform interface
Previously all command table mapping operators accepted a keyword INHERITED
defaulting to T. Thanks to that it was possible to inhibit inherited objects
mapping. Functions MAP-OVER-COMMAND-TABLE-KEYSTROKES and
MAP-OVER-COMMAND-TABLE-MENU-ITEMS were exception from this rule and the
keyword is not specified for them.

Also make the function apply-with-command-table-inheritance return NIL,
because its results were not useful.