command table: change the implementation of menus and keystrokes
This commit addresses a set of issues discovered while writing tests for the
command table accelerators (part of them was due to inconsistent specification
that is fixed in the previous commit).

- merge menu, keystroke-items and keystroke-accelerators collections into one
- rewrite mapping, find and lookup functions
- adjust add-menu-item-to-command-table and add-keystroke-to-command-table
- elements on menu that have no name are "only" keystrokes

This representation is conceptually consistent with CLIM-TOS - that was
confirmed after writing the code what only reassures me that it is "the right
thing to do" here.