tests: add tests for menu and keystroke inheritance/lookup
The specification is not consistent with regard to keystroke and lookup
treatment (most notably about keystroke inheritance and menu traversal).
Also inheritance and sub-menus are confused in one place.

These tests check for the following interpretation:

keystrokes are inherited when :inherit-menu is (member t :keystrokes)
menu items are inherited when :inherit-menu is (member t :menu)

find-keystroke-item looks for a keystroke in the command table, and if not
present and keystrokes are inherited, recursively calls the function with
ancestor

lookup-keystroke-item first calls find-keystroke-item, and if not found, it
checks in its own menus, and then maps over (potentially, depends on
:inherit-menu value) inherited submenus. The function does not look into
sub-menus that have their own keystrokes.